Term management

+++ src/main/java/com/myapp/localizationApp/repository/ProjectLanguageRepository.java
@@ -3,9 +3,10 @@ package com.myapp.localizationApp.repository;
 import com.myapp.localizationApp.entity.ProjectLanguage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ProjectLanguageRepository extends JpaRepository<ProjectLanguage, Long> {
-    List<ProjectLanguage>  findByProjectId(Long projectId);
+    List<ProjectLanguage>  findByProjectId(BigInteger projectId);
     List<ProjectLanguage> findByLanguageId(Long languageId);
 }
